Page restriction is fully functional. Non-logged-in users cannot access /admin or any of its subpages, and will be brought back to the login form. Admin can access any page, whereas managers and IT users are more restricted. Again, the logic for all of this is in permissions.js.
If a user lacks the permission, they will be redirected to the "restrict" page instead, where an error message displays and a link button that takes them back one page.

Users who already are logged will always be taken to /admin when visiting the login form.

Also implemented a logout function, for convenience.